(VariableMap, EventQueue, AlertQueue, AlertHistory): enable variable map modification even on unsaved, new objects.

+++ python/nav/models/event.py
@@ -69,12 +69,14 @@ class VariableMap(object):
 
         if hasattr(obj, self.cachename):
             return getattr(obj, self.cachename)
-        else:
+        elif obj.pk:
             variables = getattr(obj, self.variables)
             varmap = dict((var.variable, var.value)
                           for var in variables.all())
             setattr(obj, self.cachename, varmap)
             return varmap
+        else:
+            return {}
 
     def __set__(self, obj, vardict):
         if obj is None:
@@ -83,26 +85,27 @@ class VariableMap(object):
         if not hasattr(vardict, 'items'):
             raise ValueError("value must be a dict")
 
-        variables = getattr(obj, self.variables)
-        _rel_manager = getattr(obj.__class__, self.variables)
-        var_model = _rel_manager.related.model
-        related_field = _rel_manager.related.field.name
-        if vardict:
-            removed = variables.exclude(variable__in=vardict.keys())
-            removed.delete()
-        varmap = dict((v.variable, v) for v in variables.all())
-        for key, value in vardict.items():
-            if key in varmap:
-                if varmap[key].value != value:
-                    varmap[key] = value
-                    varmap[key].save()
-            else:
-                variable = var_model(**{
-                        related_field: obj,
-                        'variable': key,
-                        'value': value,
-                        })
-                variable.save()
+        if obj.pk:
+            variables = getattr(obj, self.variables)
+            _rel_manager = getattr(obj.__class__, self.variables)
+            var_model = _rel_manager.related.model
+            related_field = _rel_manager.related.field.name
+            if vardict:
+                removed = variables.exclude(variable__in=vardict.keys())
+                removed.delete()
+            varmap = dict((v.variable, v) for v in variables.all())
+            for key, value in vardict.items():
+                if key in varmap:
+                    if varmap[key].value != value:
+                        varmap[key] = value
+                        varmap[key].save()
+                else:
+                    variable = var_model(**{
+                            related_field: obj,
+                            'variable': key,
+                            'value': value,
+                            })
+                    variable.save()
 
         setattr(obj, self.cachename, vardict)
 
@@ -143,6 +146,12 @@ class EventQueue(models.Model):
             u"%s=%r" % (attr, getattr(self, attr))
             for attr in ('event_type_id', 'source_id', 'target_id', 'state'))
 
+    def save(self, *args, **kwargs):
+        new_object = self.pk is None
+        super(EventQueue, self).save(*args, **kwargs)
+        if new_object:
+            assert self.pk
+            self.varmap = self.varmap
 
 class EventType(models.Model):
     """From NAV Wiki: Defines event types."""
@@ -222,6 +231,13 @@ class AlertQueue(models.Model):
     def __unicode__(self):
         return u'Source %s, state %s, severity %d' % (
             self.source, self.get_state_display(), self.severity)
+
+    def save(self, *args, **kwargs):
+        new_object = self.pk is None
+        super(EventQueue, self).save(*args, **kwargs)
+        if new_object:
+            assert self.pk
+            self.varmap = self.varmap
 
 class AlertType(models.Model):
     """From NAV Wiki: Defines the alert types. An event type may have many alert
@@ -328,6 +344,13 @@ class AlertHistory(models.Model):
         else:
             # Stateless alert
             return None
+
+    def save(self, *args, **kwargs):
+        new_object = self.pk is None
+        super(EventQueue, self).save(*args, **kwargs)
+        if new_object:
+            assert self.pk
+            self.varmap = self.varmap
 
 class AlertHistoryMessage(models.Model):
     """From NAV Wiki: To have a history of the formatted messages too, they are
